功能增强：新增国家战略关系属性系统
核心功能：
- 新增战略关系数据库模型，支持五级关系（战争/冲突/无外交/伙伴/盟友）
- 自动初始化大国/超级大国之间的战略关系
- 在决策引擎中集成战略关系信息，LLM决策时可查看各国关系
- 提供战略关系管理API端点

数据库变更：
- 新增strategic_relationship表存储双边战略关系
- 更新simulation_project模型添加关系关联

后端变更：
- 新增StrategicRelationshipEnum服务
- 更新AgentInfo和InfoPool支持战略关系
- 仿真服务集成关系初始化和数据加载
- API路由添加战略关系查询和设置接口

前端变更：
- 优化仿真结果页面展示
- 更新统计API和数据结构

+++ app/models/action_record.py
@@ -8,7 +8,7 @@ from datetime import datetime
 from enum import Enum as PyEnum
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import Boolean, DateTime, Float, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ..models import Base
@@ -55,8 +55,8 @@ class ActionRecord(Base):
     action_category: Mapped[str] = mapped_column(String(50), nullable=False)
     action_name: Mapped[str] = mapped_column(String(255), nullable=False)
     respect_sov: Mapped[bool] = mapped_column(String(10), nullable=False, default="false")
-    initiator_power_change: Mapped[int] = mapped_column(Integer, nullable=False)
-    target_power_change: Mapped[int] = mapped_column(Integer, nullable=False)
+    initiator_power_change: Mapped[float] = mapped_column(Float, nullable=False)
+    target_power_change: Mapped[float] = mapped_column(Float, nullable=False)
     decision_detail: Mapped[str] = mapped_column(String(5000), nullable=True)
 
     created_at: Mapped[datetime] = mapped_column(
